+++ tensorcircuit/templates/__init__.py
@@ -5,5 +5,6 @@ from . import dataset
 from . import graphs
 from . import measurements
 from . import conversions
+from . import lattice
 
 costfunctions = measurements